Move DeletionConfirmationDialog class to its own .cpp file
